soundgo_api/sites/serializers.py
from rest_framework import serializers

from .models import Site


class SiteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Site
        fields = ('id', 'latitude', 'longitude', 'name', 'description')
